requirements.txt
# Core API
fastapi==0.111.0
uvicorn==0.30.1

# Database
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.18.4

# Auth & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2

# Validation
pydantic==2.7.1
email-validator==2.3.0

# Requests / JSON
requests==2.32.3
orjson==3.11.8

# Environment
python-dotenv==1.0.1

# ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
xgboost==2.0.3